Product-Composite 에서  Client 요청 종합적 처리
product-composite 에서 Client 요청을 받아 내부 모듈들의 API에 요청을 보내고 일괄적 결과 응답

+++ api/src/main/java/com/micro/api/composite/product/ProductCompositeApi.java
@@ -9,5 +9,5 @@ import com.micro.api.composite.product.dto.ProductAggregate;
 
 public interface ProductCompositeApi {
 	@GetMapping(value = "/product-composite/{productId}", produces = APPLICATION_JSON_VALUE)
-	ProductAggregate getProduct(@PathVariable int productId);
+	ProductAggregate getProduct(@PathVariable Long productId);
 }
